# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Retail Expansion Intelligence System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /competitors:
    get:
      operationId: listCompetitors
      tags: [competitors]
      summary: List competitors
      parameters:
        - name: type
          in: query
          schema: { type: ["string", "null"] }
        - name: city
          in: query
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"
    post:
      operationId: createCompetitor
      tags: [competitors]
      summary: Add a competitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
  /competitors/{id}:
    get:
      operationId: getCompetitor
      tags: [competitors]
      summary: Get competitor details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
    patch:
      operationId: updateCompetitor
      tags: [competitors]
      summary: Update competitor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitorUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
    delete:
      operationId: deleteCompetitor
      tags: [competitors]
      summary: Delete competitor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── Scoring weights ─────────────────────────────────────────
...
components:
  schemas:
...
    Competitor:
      type: object
      required: [id, name, type, city, state, createdAt]
      properties:
        id:               { type: integer }
        name:             { type: string }
        brand:            { type: ["string", "null"] }
        type:             { type: string, enum: [milma_booth, juice_shop, tea_shop, bubble_tea, cafe, fresh_juice_chain, ice_cream_chain, convenience_store, mall_food_court, petrol_station, other] }
        city:             { type: string }
        state:            { type: string }
        latitude:         { type: ["number", "null"] }
        longitude:        { type: ["number", "null"] }
        address:          { type: ["string", "null"] }
        priceRange:       { type: ["string", "null"], enum: [budget, mid, premium, null] }
        avgRating:        { type: ["number", "null"] }
        reviewCount:      { type: ["integer", "null"] }
        deliveryAvailable: { type: ["boolean", "null"] }
        operatingHours:   { type: ["string", "null"] }
        strengths:        { type: ["string", "null"] }
        weaknesses:       { type: ["string", "null"] }
        marketShare:      { type: ["number", "null"] }
        notes:            { type: ["string", "null"] }
        createdAt:        { type: string }
    CompetitorInput:
      type: object
      required: [name, type, city, state]
      properties:
        name:             { type: string }
        brand:            { type: string }
        type:             { type: string }
        city:             { type: string }
        state:            { type: string }
        latitude:         { type: number }
        longitude:        { type: number }
        address:          { type: string }
        priceRange:       { type: string }
        avgRating:        { type: number }
        reviewCount:      { type: integer }
        deliveryAvailable: { type: boolean }
        operatingHours:   { type: string }
        strengths:        { type: string }
        weaknesses:       { type: string }
        marketShare:      { type: number }
        notes:            { type: string }
    CompetitorUpdate:
      type: object
      properties:
        name:             { type: string }
        brand:            { type: string }
        priceRange:       { type: string }
        avgRating:        { type: number }
        reviewCount:      { type: integer }
        deliveryAvailable: { type: boolean }
        operatingHours:   { type: string }
        strengths:        { type: string }
        weaknesses:       { type: string }
        marketShare:      { type: number }
        notes:            { type: string }

    # ─── Demand Profile ────────────────────────────────────────
...